Add udpPort function
Probably not elegant enough to be accepted but I'd like a way of finding
out the port number for a UDP connection.  If I pass a port number of 0
to udpServer I get a number allocated for me, but afaict I can't then find 
out what it is, because the UDP data constructor hides the socket from me
and isn't exported.